fix: move error definitions where they are used
As we have distinct errors, move them where they are used.
