+++ src/serde/mod.rs
@@ -8,6 +8,30 @@ mod ser;
 pub use de::from_ipld;
 pub use ser::{to_ipld, Serializer};
 
+/// Error during Serde operations.
+#[derive(Clone, Debug)]
+pub struct SerdeError(String);
+
+impl fmt::Display for SerdeError {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "serde error: {}", self.0)
+    }
+}
+
+impl serde::de::Error for SerdeError {
+    fn custom<T: core::fmt::Display>(msg: T) -> Self {
+        Self(msg.to_string())
+    }
+}
+
+impl serde::ser::Error for SerdeError {
+    fn custom<T: fmt::Display>(msg: T) -> Self {
+        Self(msg.to_string())
+    }
+}
+
+impl serde::ser::StdError for SerdeError {}
+
 #[cfg(test)]
 mod tests {
     use std::collections::BTreeMap;
